chore: refactor to hook
